refactor: move OpEnergy related database migration code into separate module in order to not to interfere with mempool's migration code too much
